requirements.txt
python-telegram-bot==21.6
google-genai>=0.5.0
Flask
APScheduler
